add movie crud functions

+++ src/main/java/net/kleditzsch/SmartHome/model/movie/movie/MovieSeries.java
@@ -30,7 +30,7 @@ public class MovieSeries extends Element {
     /**
      * Filme der Filmreihe
      */
-    private SortedSet<SeriesMovie> movies = new TreeSet<>(Comparator.comparingInt(SeriesMovie::getSeriesOrder));
+    private SortedSet<SeriesMovie> movies = new TreeSet<>(Comparator.comparingDouble(SeriesMovie::getSeriesOrder));
 
     /**
      * gibt den Titel zurück
